+++ src/components/Sidebar.vue
@@ -60,7 +60,13 @@
           </div>
           <!-- <transition-group v-if="tab_group.open && show_tabs && ! isTabGroupDragSource( tab_group )" :class="[ `sidebar-tab-group-tabs-list--${ theme }` ]" tag="div" :name="`sidebar-tab-group-tabs-list--${ theme }__item--transition`"> -->
           <div v-if="tab_group.open && show_tabs && tab_group.tabs.length" :class="[ bem( 'list-flex-col', { theme, 'is-dragging': drag_state.is_dragging, 'is-searching': is_searching, 'is-max-height': true } ) ]">
-            <div :class="bem( `list-flex-col__item`, { 'active': tab.active, 'drag-target-index': drag_state.target.tab_id === tab.id && ! isSelected( tab ), 'drag-selected': isSelected( tab ), 'drag-source': drag_state.is_dragging && isSelected( tab ), 'search': getTabSearchState( tab ) } )"
+            <div :class="bem( `list-flex-col__item`, {
+                  'active': tab.active,
+                  'drag-target-index': drag_state.target.tab_id === tab.id && ! isSelected( tab ),
+                  'drag-selected': isSelected( tab ) && ( ! drag_state.is_dragging || drag_state.source.type === 'tab' ),
+                  'drag-source': drag_state.is_dragging && drag_state.source.type === 'tab' && isSelected( tab ),
+                  'search': getTabSearchState( tab )
+                } )"
                 v-for="tab in tab_group.tabs" :key="tab.id" :tab="tab"
                 :title="tab.title"
                 @click.ctrl="toggleTabSelection( tab )" @click.shift="toggleTabBatchSelection( tab )" @click.middle.exact="closeTab( tab )" @click.exact="openTab( tab.id )"