+++ src/components/Sidebar.vue
@@ -60,7 +60,13 @@
           </div>
           <!-- <transition-group v-if="tab_group.open && show_tabs && ! isTabGroupDragSource( tab_group )" :class="[ `sidebar-tab-group-tabs-list--${ theme }` ]" tag="div" :name="`sidebar-tab-group-tabs-list--${ theme }__item--transition`"> -->
           <div v-if="tab_group.open && show_tabs && tab_group.tabs.length" :class="[ bem( 'list-flex-col', { theme, 'is-dragging': drag_state.is_dragging, 'is-searching': is_searching, 'is-max-height': true } ) ]">
-            <div :class="bem( `list-flex-col__item`, { 'active': tab.active, 'drag-target-index': drag_state.target.tab_id === tab.id && ! isSelected( tab ), 'drag-selected': isSelected( tab ), 'drag-source': drag_state.is_dragging && isSelected( tab ), 'search': getTabSearchState( tab ) } )"
+            <div :class="bem( `list-flex-col__item`, {
+                  'active': tab.active,
+                  'drag-target-index': drag_state.target.tab_id === tab.id && ! isSelected( tab ),
+                  'drag-selected': isSelected( tab ) && ( ! drag_state.is_dragging || drag_state.source.type === 'tab' ),
+                  'drag-source': drag_state.is_dragging && drag_state.source.type === 'tab' && isSelected( tab ),
+                  'search': getTabSearchState( tab )
+                } )"
                 v-for="tab in tab_group.tabs" :key="tab.id" :tab="tab"
                 :title="tab.title"
                 @click.ctrl="toggleTabSelection( tab )" @click.shift="toggleTabBatchSelection( tab )" @click.middle.exact="closeTab( tab )" @click.exact="openTab( tab.id )"
@@ -116,15 +122,16 @@
     </div>
     <!-- </transition-group> -->
     <div :class="bem( `empty-dropzone--${ theme }`, { 'drag-tab-target': drag_state.source.type === 'tab' && drag_state.target.tab_group_new, 'drag-tab_group-target': drag_state.source.type === 'tab_group' && drag_state.target.tab_group_new } )"
+        ref="empty_dropzone"
         @click.right.prevent
-        @dragenter="onTabGroupDragEnter( $event, null, null, 'dropzone' )"
+        @dragenter="updateDropzoneHeight() || onTabGroupDragEnter( $event, null, null, 'dropzone' )"
         @dragover.prevent
         @dragleave="onTabGroupDragLeave"
         @dragend="onTabDragEnd"
         @drop="onTabGroupDrop( $event, null, null, 'dropzone' )"
     >
       <div :class="[ `empty-dropzone--${ theme }__drag-target-ink` ]"></div>
-      <svg v-if="drag_state.target.tab_group_new && drag_state.source.type === 'tab'" :class="[ `empty-dropzone--${ theme }__drag-target-icon` ]" xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 16 16">
+      <svg v-if="show_dropzone_icon" :class="[ `empty-dropzone--${ theme }__drag-target-icon` ]" xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 16 16">
         <path d="M14 7H9V2a1 1 0 0 0-2 0v5H2a1 1 0 1 0 0 2h5v5a1 1 0 0 0 2 0V9h5a1 1 0 0 0 0-2z"></path>
       </svg>
     </div>
@@ -234,6 +241,7 @@ export default {
       search_matched_tab_ids: [],
       search_missed_tab_ids: [],
       selected_tab_ids: [],
+      show_dropzone_icon: false,
       show_header: true,
       show_tabs: true,
       show_tabs_count: true,
@@ -487,6 +495,16 @@ export default {
       toggleTabBatchSelection.call( this, tab.id )
       const highlighted_tab_ids = this.selected_tab_ids.slice( 0, this.selected_tab_ids.length )
       window.background.setHighlightedTabIds( window.store, this.window_id, highlighted_tab_ids )
+    },
+    updateDropzoneHeight() {
+      // Only display the icon if the dropzone is bigger than 35 height
+      if( ! this.$refs.empty_dropzone ) {
+        this.show_dropzone_icon = false
+      } else if( this.$refs.empty_dropzone.clientHeight < 36 ) {
+        this.show_dropzone_icon = false
+      } else {
+        this.show_dropzone_icon = true
+      }
     },
   }
 }
@@ -924,7 +942,7 @@ $tab-groups-list-item-header__themes: (
   align-items: center;
   position: sticky;
   top: 0;
-  z-index: 1;
+  z-index: 2;
   border-bottom: $light-border-color 1px solid;
   background-color: map-get( $colors, --background-color );
   cursor: pointer;