+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - new group icon display edge case
 - active tab display in blue
 - drag styling on hilighted tabs while group is moving
+- flash of drag styling on active tab switch
 
 ## [0.0.13] - 2019-07-21